Change user schema to have name and removed name, email, phone from address

+++ controller/Cart.js
@@ -10,9 +10,21 @@ exports.addToCart = async function (req, res) {
 		const doc = await cartItem.save();
 		const item = await Cart.findById(doc._id).populate('product').populate('user');
 
-		res.status(200).json(item);
+		res.status(200).json({
+			success: true,
+			message: `${item.product.title} added to cart successfully`,
+			data: {item},
+			error: null
+		});
+
+
 	} catch (err) {
-		res.status(400).json(err)
+		res.status(400).json({
+			success: false,
+			message: err.message,
+			data: null,
+			error: err
+		})
 	}
 };
 
@@ -23,9 +35,19 @@ exports.fetchCartItemsByUserId = async function (req, res) {
 	try {
 		const cart = await Cart.find({user: id}).populate('product').populate('user');
 
-		res.status(200).json(cart);
+		res.status(200).json({
+			success: true,
+			message: 'Cart items fetched successfully',
+			data: {cart},
+			error: null
+		});
 	} catch (err) {
-		res.status(400).json(err)
+		res.status(400).json({
+			success: false,
+			message: err.message,
+			data: null,
+			error: err
+		})
 	}
 };
 
@@ -34,11 +56,36 @@ exports.updateCartItem = async function (req, res) {
 	const { id } = req.params;
 
 	try {
-		const cart = await Cart.findByIdAndUpdate(id, req.body, { new: true }).populate('product').populate('user');
+		const cartItem = await Cart.findByIdAndUpdate(
+			id,
+			req.body,
+			{ new: true }
+		).populate('product').populate('user');
 
-		res.status(200).json(cart);
+		if (!cartItem) {
+			return res.status(404).json({
+				success: false,
+				message: 'Unable to update item which is not added in cart',
+				data: null,
+				error: {
+					message: `cart item with id ${id} not found`
+				}
+			});
+		}
+
+		res.status(200).json({
+			success: true,
+			message: 'cart item updated successfully',
+			data: { item: cartItem },
+			error: null
+		});
 	} catch (err) {
-		res.status(400).json(err)
+		res.status(400).json({
+			success: false,
+			message: err.message,
+			data: null,
+			error: err
+		})
 	}
 };
 
@@ -48,11 +95,32 @@ exports.deleteCartItem = async function (req, res) {
 
 	try {
 		const cartItem = await Cart.findById(id).populate('product', 'id title');
-		await Cart.findByIdAndDelete(id);
 
-		res.status(200).json(cartItem);
+		if (!cartItem) {
+			return res.status(404).json({
+				success: false,
+				message: 'Unable to delete item which is not added in cart',
+				data: null,
+				error: {
+					message: `cart item with id ${id} not found`
+				}
+			});
+		}
+
+		await Cart.findByIdAndDelete(id);
+		res.status(200).json({
+			success: true,
+			message: 'Cart item deleted successfully',
+			data: { item: cartItem },
+			error: null
+		});
 	} catch (err) {
-		res.status(400).json(err)
+		res.status(400).json({
+			success: false,
+			message: err.message,
+			data: null,
+			error
+		})
 	}
 };
 
@@ -65,8 +133,18 @@ exports.clearCart = async function (req, res) {
 			await Cart.findByIdAndDelete(id);
 		}
 
-		res.status(200).json({success: true, message: 'Cart Cleared Successfully'});
+		res.status(200).json({
+			success: true,
+			message: 'Cart Cleared Successfully',
+			data: null,
+			error: null
+		});
 	} catch (e) {
-		res.status(400).json(e);
+		res.status(400).json({
+			success: false,
+			message: e.message,
+			data: null,
+			error: e
+		});
 	}
 }
